refactor: updated the project name
Remove the 0 from the name as it is out of place. No the name can be
more understandable
