+++ backend/src/main/java/com/fake_orgasm/Application.java
@@ -1,4 +1,4 @@
-package com.fake_orgasm0;
+package com.fake_orgasm;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
